backend/requirements/base.txt
# Core Django
Django>=5.0,<5.1
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9

# REST Framework
djangorestframework>=3.14.0
django-filter>=23.5
drf-spectacular>=0.27.0

# Authentication
django-allauth>=0.59.0

# CORS headers
django-cors-headers>=4.3.0

# Environment variables
dj-database-url>=2.1.0

# Timezone support
pytz>=2024.1
